+++ frontend/app/src/components/form/generic-selector.tsx
@@ -14,8 +14,8 @@ import { useSchema } from "@/hooks/useSchema";
 import LoadingScreen from "@/screens/loading-screen/loading-screen";
 import { getObjectPermissionsQuery } from "@/screens/permission/queries/getObjectPermissions";
 import { PermissionData } from "@/screens/permission/types";
+import { getPermission } from "@/screens/permission/utils";
 import { genericsState, profilesAtom, schemaState } from "@/state/atoms/schema.atom";
-import { getPermission } from "@/utils/permissions";
 import { gql } from "@apollo/client";
 import { useAtomValue } from "jotai/index";
 import React, { useId, useState } from "react";
