+++ frontend/app/src/hooks/useObjectDetails.ts
@@ -1,10 +1,10 @@
 import { PROFILE_KIND, TASK_OBJECT } from "@/config/constants";
 import { getObjectDetailsPaginated } from "@/graphql/queries/objects/getObjectDetails";
 import useQuery from "@/hooks/useQuery";
+import { getPermission } from "@/screens/permission/utils";
 import { IModelSchema, genericsState } from "@/state/atoms/schema.atom";
 import { isGeneric } from "@/utils/common";
 import { getSchemaObjectColumns, getTabs } from "@/utils/getSchemaObjectColumns";
-import { getPermission } from "@/utils/permissions";
 import { gql } from "@apollo/client";
 import { useAtomValue } from "jotai";
